# Conflicts:
#	BrawlCrate/BrawlCrate.csproj
#	BrawlCrate/NodeWrappers/Animations/CHR0Wrapper.cs
#	BrawlCrate/NodeWrappers/Archives/ARCWrapper.cs
#	BrawlCrate/NodeWrappers/BrawlEx/RSTCWrapper.cs
#	BrawlCrate/NodeWrappers/Graphics/TEX0Wrapper.cs
#	BrawlCrate/NodeWrappers/MDL0/MDL0Wrapper.cs
#	BrawlCrate/Properties/Settings.Designer.cs
#	BrawlLib/BrawlLib.csproj
#	CanaryBuild/Canary

+++ BrawlLib/SSBB/ResourceNodes/Modules/RELSectionNode.cs
@@ -25,7 +25,7 @@ namespace BrawlLib.SSBB.ResourceNodes
         public bool _isCodeSection = false;
         public bool _isBSSSection = false;
         public int _dataOffset = 0;
-        public int _endBufferSize = 0x0;
+        public uint _endBufferSize = 0x0;
         public uint _dataSize;
         public int _dataAlign;
 
@@ -84,7 +84,7 @@ namespace BrawlLib.SSBB.ResourceNodes
         }
         public override int OnCalculateSize(bool force, bool rebuilding = true)
         {
-            return _dataBuffer.Length + _endBufferSize;
+            return _dataBuffer.Length + (int)_endBufferSize;
         }
 
         public override void OnRebuild(VoidPtr address, int length, bool force)
